adding rule for Style/IfUnlessModifier

+++ spec/rails_helper.rb
@@ -151,8 +151,6 @@ RSpec.configure do |config|
   # clean up carrierwave uploads
   # https://github.com/carrierwaveuploader/carrierwave/wiki/How-to:-Cleanup-after-your-Rspec-tests
   config.after(:all) do
-    if Rails.env.test?
-      FileUtils.rm_rf(Dir[Rails.root.join('spec', 'support', 'uploads')])
-    end
+    FileUtils.rm_rf(Dir[Rails.root.join('spec', 'support', 'uploads')]) if Rails.env.test?
   end
 end
